// src/types/enum.ts
export enum POSITION {
	OH = 'OH',
	OP = 'OP',
	MB = 'MB',
	S = 'S',
	L = 'L',
}

export enum ACTION {
	ACE = 'ACE',
	KILL = 'KILL',
	BLOCK = 'BLOCK',
	SERVICE_NOT_IN = 'SERVICE_NOT_IN',
	SERVICE_OUTBALL = 'SERVICE_OUTBALL',
	SPIKE_NOT_IN = 'SPIKE_NOT_IN',
	SPIKE_OUTBALL = 'SPIKE_OUTBALL',
	TOUCH_OUT = 'TOUCH_OUT',
	TIP = 'TIP',
}

export enum VALIDATION {
	NET_TOUCH = 'NET_TOUCH',
	FOOT_FAULT = 'FOOT_FAULT',
	DOUBLE_CONTACT = 'DOUBLE_CONTACT',
	CATCH = 'CATCH',
	CENTER_LINE_FAULT = 'CENTER_LINE_FAULT',
	POSITION_FAULT = 'POSITION_FAULT',
	DELAY_IN_SERVICE = 'DELAY_IN_SERVICE',
	FOUR_HITS = 'FOUR_HITS',
	REACHING_BEYOND_THE_NET = 'REACHING_BEYOND_THE_NET',
	ATTACK_HIT_FAULT = 'ATTACK_HIT_FAULT',
}